move recorder to components
